+++ packages/stream_chat_localizations/CHANGELOG.md
@@ -1,6 +1,10 @@
+## Upcoming
+
+* Added support for [Spanish](https://github.com/GetStream/stream-chat-flutter/blob/master/packages/stream_chat_localizations/lib/src/stream_chat_localizations_es.dart) locale.
+
 ## 1.0.2
 
-* Updated stream_chat_flutter dependency
+* Updated `stream_chat_flutter` dependency
 
 ## 1.0.1
 
@@ -8,4 +12,8 @@
 
 ## 1.0.0
 
-* First release
+* Initial Release with support for 4 locales
+    - [English](https://github.com/GetStream/stream-chat-flutter/blob/master/packages/stream_chat_localizations/lib/src/stream_chat_localizations_en.dart)
+    - [Hindi](https://github.com/GetStream/stream-chat-flutter/blob/master/packages/stream_chat_localizations/lib/src/stream_chat_localizations_hi.dart)
+    - [Italian](https://github.com/GetStream/stream-chat-flutter/blob/master/packages/stream_chat_localizations/lib/src/stream_chat_localizations_it.dart)
+    - [French](https://github.com/GetStream/stream-chat-flutter/blob/master/packages/stream_chat_localizations/lib/src/stream_chat_localizations_fr.dart)